generic to check formula and data only
It helps define non-standard evaluation for X and Y - since they don't need to be evalauted before method dispatch - but the drawback is that backward compatibility might become compromised, as if a user has an existing code of, say, pls(matrix1, matrix2) with unnamed arguments, the new generic will evaluate them as formula and data and create error. There is a workaround of definig methods for numeric first 2 arguments (formula and data in generic, but X and Y in old function) and then changing them to X and Y in that case with a message so they know such a shift happened for code diagnostics.